+++ demos/compose/src/main/java/androidx/media3/demo/compose/MainActivity.kt
@@ -19,15 +19,18 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.material3.Surface
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.size
+import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.unit.dp
 import androidx.media3.common.MediaItem
 import androidx.media3.common.Player
+import androidx.media3.demo.compose.buttons.PlayPauseButton
 import androidx.media3.demo.compose.data.videos
 import androidx.media3.exoplayer.ExoPlayer
 import androidx.media3.ui.compose.PlayerSurface
@@ -39,27 +42,24 @@ class MainActivity : ComponentActivity() {
     super.onCreate(savedInstanceState)
     enableEdgeToEdge()
     setContent {
-      Surface {
-        Column(
-          verticalArrangement = Arrangement.Center,
-          horizontalAlignment = Alignment.CenterHorizontally,
-        ) {
-          val context = LocalContext.current
-          val exoPlayer = remember {
-            ExoPlayer.Builder(context).build().apply {
-              setMediaItem(MediaItem.fromUri(videos[0]))
-              prepare()
-              playWhenReady = true
-              repeatMode = Player.REPEAT_MODE_ONE
-            }
-          }
-          PlayerSurface(
-            player = exoPlayer,
-            surfaceType = SURFACE_TYPE_SURFACE_VIEW,
-            modifier = Modifier.align(Alignment.CenterHorizontally),
-          )
+      val context = LocalContext.current
+      val exoPlayer = remember {
+        ExoPlayer.Builder(context).build().apply {
+          setMediaItem(MediaItem.fromUri(videos[0]))
+          prepare()
+          playWhenReady = true
+          repeatMode = Player.REPEAT_MODE_ONE
         }
       }
+      MediaPlayerScreen(player = exoPlayer, modifier = Modifier.fillMaxSize())
+    }
+  }
+
+  @Composable
+  private fun MediaPlayerScreen(player: Player, modifier: Modifier = Modifier) {
+    Box(modifier) {
+      PlayerSurface(player = player, surfaceType = SURFACE_TYPE_SURFACE_VIEW)
+      PlayPauseButton(player, Modifier.align(Alignment.Center).size(100.dp))
     }
   }
 }