+++ RELEASENOTES.md
@@ -57,6 +57,8 @@
 *   IMA extension:
 *   UI:
     *   Add `PlayerSurface` Composable to `media3-ui-compose` module.
+    *   Add `PlayPauseButtonState` class and `rememberPlayPauseButtonState`
+        Composable to `media3-ui-compose` module.
 *   Downloads:
 *   OkHttp Extension:
 *   Cronet Extension:
@@ -74,6 +76,8 @@
 *   Cast Extension:
 *   Test Utilities:
 *   Demo app:
+    *   Add `PlayPauseButton` Composable UI element to `demo-compose` utilizing
+        `PlayPauseButtonState`.
 *   Remove deprecated symbols:
     *   Remove deprecated `AudioMixer.create()` method. Use
         `DefaultAudioMixer.Factory().create()` instead.